refactor(project-structure): implement feature-based architecture
- Reorganize components, types, and utils related to articles and authors into feature folders
- Move shared components and icons to shared folder
- Update import paths across all affected files
- Rename some files and constants for better clarity and consistency
