--- src/lib/data/authors.json
@@ -1,46 +1,0 @@
-[
-  {
-    "userId": "1",
-    "role": "AUTHOR",
-    "displayName": "Alex Johnson",
-    "bio": "Full-stack developer with a passion for creating scalable web applications. Specializing in modern JavaScript frameworks and cloud technologies.",
-    "profileImage": null,
-    "createdAt": "2023-01-15T00:00:00.000Z",
-    "updatedAt": "2023-06-20T00:00:00.000Z",
-    "articleCount": 24,
-    "subscriberCount": 310
-  },
-  {
-    "userId": "2",
-    "role": "AUTHOR",
-    "displayName": "Maria Garcia",
-    "bio": "UI/UX designer and frontend developer focused on creating accessible and performant web experiences.",
-    "profileImage": null,
-    "createdAt": "2023-03-10T00:00:00.000Z",
-    "updatedAt": "2023-07-12T00:00:00.000Z",
-    "articleCount": 12,
-    "subscriberCount": 150
-  },
-  {
-    "userId": "3",
-    "role": "AUTHOR",
-    "displayName": "Samuel Chen",
-    "bio": "Data scientist and technical writer. Exploring the intersection of machine learning and web development.",
-    "profileImage": null,
-    "createdAt": "2022-11-05T00:00:00.000Z",
-    "updatedAt": "2023-08-01T00:00:00.000Z",
-    "articleCount": 31,
-    "subscriberCount": 980
-  },
-  {
-    "userId": "4",
-    "role": "AUTHOR",
-    "displayName": "Elena Rodriguez",
-    "bio": "Mobile development expert with a focus on cross-platform solutions. Passionate about creating intuitive user experiences.",
-    "profileImage": null,
-    "createdAt": "2023-05-22T00:00:00.000Z",
-    "updatedAt": "2023-08-15T00:00:00.000Z",
-    "articleCount": 18,
-    "subscriberCount": 420
-  }
-]
