+++ src/lib/features/admin/components/AdminSidebar.svelte
