+++ src/lib/features/auth/components/SuccessMessage.svelte
