+++ src/lib/features/auth/utils/validate.ts
