+++ src/lib/features/user/components/AuthorSidebar.svelte
