+++ src/lib/shared/components/navbar/Navbar.svelte
