+++ src/lib/shared/icons/BookIcon.svelte
