+++ src/lib/shared/icons/FollowerIcon.svelte
