+++ src/lib/shared/icons/HeartIcon.svelte
