+++ src/lib/shared/icons/NextIcon.svelte
