+++ src/lib/shared/icons/PreviousIcon.svelte
