+++ src/lib/shared/icons/SearchIcon.svelte
