+++ src/lib/shared/utils/common.ts
