+++ src/routes/+layout.server.ts
@@ -1,6 +1,6 @@
 import { redirect } from '@sveltejs/kit';
 import type { LayoutServerLoad } from './$types';
-import { PUBLIC_ROUTES } from '$lib/constants/auth';
+import { PUBLIC_ROUTES } from '$lib/features/auth/constants/auth';
 
 export const load: LayoutServerLoad = async ({ locals, cookies, url }) => {
 	const isPublic = PUBLIC_ROUTES.some((rx) => rx.test(url.pathname));
