+++ src/routes/admin/+layout.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-	import AdminSidebar from '$lib/components/admin/AdminSidebar.svelte';
+	import AdminSidebar from '$lib/features/admin/components/AdminSidebar.svelte';
 
 	const props = $props();
